requirements.txt
# 基础依赖
requests==2.31.0
beautifulsoup4==4.12.2
numpy>=1.24.0

# Selenium相关
selenium>=4.15.0
webdriver-manager>=4.0.1

# OpenCV
opencv-python>=4.8.0

# 日期时间处理
python-dateutil>=2.8.2

# 图像处理
Pillow>=10.0.0

# 日志处理
loguru>=0.7.0

# 网络请求
urllib3>=2.0.7

alembic==1.13.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.1.0
Babel==2.15.0
blinker==1.8.0
bs4==0.0.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crcmod==1.7
cryptography==41.0.7
Deprecated==1.2.14
distro==1.9.0
EbookLib==0.18
et_xmlfile==2.0.0
flasgger==0.9.7.1
Flask==3.0.0
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-JWT-Extended==4.5.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==4.9.3
Mako==1.3.9
MarkupSafe==3.0.2
minio==7.2.15
mistune==3.1.2
openpyxl==3.1.5
oss2==2.19.1
packaging==24.1
pandas==2.2.2
protobuf==4.25.3
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.38
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
